starletterProject/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==4.2.14
django-allauth==0.63.6
django-cors-headers==4.4.0
django-multiselectfield==0.1.13
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
idna==3.7
pillow==10.4.0
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
